refactor: reorganize imports across CLI and config modules for improved clarity and structure
