+++ src/cli/commands/history.command.ts
@@ -3,14 +3,13 @@ import {select} from '@inquirer/prompts';
 import chalk from 'chalk';
 import Table from 'cli-table3';
 import {Command} from 'commander';
+import {DatabaseManager} from '../../core/database/manager.js';
 import {
-  type ChatHistoryOptions,
-  DatabaseManager,
   DEFAULT_HISTORY_LIMIT,
   DEFAULT_PROFILE,
-  formatRelativeTime,
-  truncateText,
-} from '../../core/index.js';
+} from '../../core/shared/constants.js';
+import type {ChatHistoryOptions} from '../../core/shared/types.js';
+import {formatRelativeTime, truncateText} from '../../core/shared/utils.js';
 
 export const createHistoryCommand = (): Command => {
   const historyCommand = new Command('history');
