+++ src/core/configs/managers/display.manager.ts
@@ -1,7 +1,7 @@
 import chalk from 'chalk';
 import {getProvider} from '../../providers/index.js';
 import {maskApiKey, type RawiCredentials} from '../../shared/index.js';
-import type {IConfigDisplayManager} from '../index.js';
+import type {IConfigDisplayManager} from '../interfaces/config-manager.interface.js';
 
 export class ConfigDisplayManager implements IConfigDisplayManager {
   displayCredentials(credentials: RawiCredentials): void {
