+++ src/core/templates/utils.ts
@@ -1,7 +1,7 @@
 import {select} from '@inquirer/prompts';
 import chalk from 'chalk';
 import Table from 'cli-table3';
-import {type ActTemplate, actTemplates} from './index.js';
+import {type ActTemplate, actTemplates} from './act.js';
 
 export const findActTemplate = (id: string): ActTemplate | undefined => {
   return actTemplates.find((template) => template.id === id);
